Css tinkering
Played around with the CSS properties for the webpage, including changing the colors of the TH data to violet and the TD data to Ivory to make legible on the grey background. as well as changed the Link colors to a light shade of blue from a dark shade of purple.